Add files via upload
H0N3Y_SC4N was a project for one of my Cybersecurity classes. 
Its purpose was to simulate open and closed ports to protect systems from port scan attacks. 
Additionally, it creates a log file with all the requests made to the host system.
The script features a minimal interface allowing users to choose the range of ports that it will modify packets for.